+++ Seq2SeqSharp/Corpus/VisionTextCorpus.cs
@@ -49,9 +49,9 @@ namespace Seq2SeqSharp.Corpus
         private string m_sortedIndexedDataSetFilePath = "";
         private int m_batchNumInTotal = 0;
 
-        public List<Dictionary<string, int>> CountTokenFreqs()
+        public List<Dictionary<string, long>> CountTokenFreqs()
         {
-            List<Dictionary<string, int>> td = new List<Dictionary<string, int>>();
+            List<Dictionary<string, long>> td = new List<Dictionary<string, long>>();
 
             for (int i = 0; i < m_tgtFileList.Count; i++)
             {
@@ -77,7 +77,7 @@ namespace Seq2SeqSharp.Corpus
                     {
                         for (int j = 0; j < tgtGroups.Length; j++)
                         {
-                            td.Add(new Dictionary<string, int>());
+                            td.Add(new Dictionary<string, long>());
                         }
                     }
